+++ tutorial/05_kallisto.sh
@@ -13,8 +13,8 @@ NAME=(
     "HEK293_siCTCF_rep2"
 )
 
-sing="singularity exec --bind /work,/work2,/work3 /work3/SingularityImages/rumball.0.5.2.sif"
-#sing="singularity exec rumball.sif"
+#sing="singularity exec --bind /work,/work2,/work3 /work3/SingularityImages/rumball.0.6.0.sif"
+sing="singularity exec rumball.sif"
 
 Ddir=Ensembl-GRCh38/
 
@@ -25,7 +25,7 @@ do
     fq1=fastq/${ID[$i]}_1.fastq.gz
     fq2=fastq/${ID[$i]}_2.fastq.gz
     index=$Ddir/kallisto-indexes/genome
-#    $sing kallisto.sh -p 64 ${NAME[$i]} "$fq1 $fq2" $Ddir reverse
+    $sing kallisto.sh -p 64 ${NAME[$i]} "$fq1 $fq2" $Ddir reverse
 done
 
 s=""
@@ -34,16 +34,12 @@ do
     s="$s kallisto/${NAME[$i]}/abundance.tsv"
 done
 
-
-Ctrl="bowtie2/HEK293_Control_rep1 bowtie2/HEK293_Control_rep2"
-siCTCF="bowtie2/HEK293_siCTCF_rep1 bowtie2/HEK293_siCTCF_rep2"
-
 # For DESeq2
 mkdir -p Matrix_deseq2_kallisto
 $sing kallisto_merge.sh "$s" Matrix_deseq2_kallisto/HEK293 $Ddir
 $sing DESeq2.sh -k Matrix_deseq2_kallisto/HEK293 2:2 Control:siCTCF Human
 
 # For edgeR
-#mkdir -p Matrix_edgeR_kallisto
-#rsem_merge.sh "$Ctrl $siCTCF" Matrix_edgeR_kallisto/HEK293 $Ddir
-#edgeR.sh Matrix_edgeR_kallisto/HEK293 2:2 Control:siCTCF Human
+mkdir -p Matrix_edgeR_kallisto
+$sing kallisto_merge.sh "$s" Matrix_edgeR_kallisto/HEK293 $Ddir
+$sing edgeR.sh -k Matrix_edgeR_kallisto/HEK293 2:2 Control:siCTCF Human